+++ app/src/main/res/layout/rv_catering.xml
@@ -17,7 +17,7 @@
         android:textSize="39sp" />
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycle_restaurants"
+        android:id="@+id/recycle_catering"
         android:layout_width="match_parent"
         android:layout_height="660dp"
         android:layout_marginStart="2dp"
